% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_to_hpa.R
\name{bar_to_hpa}
\alias{bar_to_hpa}
\title{Bar to hectopascal conversion}
\usage{
bar_to_hpa(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of pressure 
    from bars to hectopascals
}
\examples{
bar_to_hpa(99)
bar_to_hpa(1033)
bar_to_hpa(687)
bar_to_hpa(c(350, 8332, 11033))
}
