% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_to_pa.R
\name{bar_to_pa}
\alias{bar_to_pa}
\title{Bar to pascal conversion}
\usage{
bar_to_pa(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of pressure 
    from bars to pascals
}
\examples{
bar_to_pa(939)
bar_to_pa(10233)
bar_to_pa(637)
bar_to_pa(c(3350, 820, 110333))
}
