% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byte_to_bit.R
\name{byte_to_bit}
\alias{byte_to_bit}
\title{Byte to bit conversion}
\usage{
byte_to_bit(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of data size 
    from bytes to bits
}
\examples{
byte_to_bit(99)
byte_to_bit(1033)
byte_to_bit(687)
byte_to_bit(c(350, 8332, 11033))
}
