% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_to_f.R
\name{c_to_f}
\alias{c_to_f}
\title{Celsius to fahrenheit conversion}
\usage{
c_to_f(x)
}
\arguments{
\item{x}{a vector containing a value in celsius}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of temperature 
    from celsius' to fahrenheits
}
\examples{
c_to_f(75)
c_to_f(110)
c_to_f(0)
c_to_f(c(0, 75, 110))
}
