% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_to_k.R
\name{c_to_k}
\alias{c_to_k}
\title{Celsius to kelvin conversion}
\usage{
c_to_k(x)
}
\arguments{
\item{x}{a vector containing a value in celsius}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of temperature 
    from celsius' to kelvins
}
\examples{
c_to_k(75)
c_to_k(110)
c_to_k(0)
c_to_k(c(0, 75, 110))
}
