% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_to_c.R
\name{f_to_c}
\alias{f_to_c}
\title{Fahrenheit to celsius conversion}
\usage{
f_to_c(x)
}
\arguments{
\item{x}{a vector containing a value in fahrenheit}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of temperature 
    from fahrenheits to celsius'
}
\examples{
f_to_c(75)
f_to_c(110)
f_to_c(0)
f_to_c(c(0, 75, 110))
}
