% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_to_k.R
\name{f_to_k}
\alias{f_to_k}
\title{Fahrenheit to kelvin conversion}
\usage{
f_to_k(x)
}
\arguments{
\item{x}{a vector containing a value in fahrenheit}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of temperature 
    from fahrenheits to kelvins
}
\examples{
f_to_k(75)
f_to_k(110)
f_to_k(0)
f_to_k(c(0, 75, 110))
}
