% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_to_kg.R
\name{g_to_kg}
\alias{g_to_kg}
\title{Gram to kilogram conversion}
\usage{
g_to_kg(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of weight
    from gram to kilogram
}
\examples{
g_to_kg(99)
g_to_kg(1033)
g_to_kg(687)
g_to_kg(c(350, 8332, 11033))
}
