% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb_to_mb.R
\name{gb_to_mb}
\alias{gb_to_mb}
\title{Gigabyte to megabyte conversion}
\usage{
gb_to_mb(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of data
    from gigabytes to megabytes
}
\examples{
gb_to_mb(99)
gb_to_mb(1033)
gb_to_mb(687)
gb_to_mb(c(350, 8332, 11033))
}
