% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpa_to_pa.R
\name{hpa_to_pa}
\alias{hpa_to_pa}
\title{Hectopascal to pascal conversion}
\usage{
hpa_to_pa(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of pressure 
    from hectopascals to pascals
}
\examples{
hpa_to_pa(9929)
hpa_to_pa(10213)
hpa_to_pa(6637)
hpa_to_pa(c(503, 8100, 11160))
}
