% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_to_f.R
\name{k_to_f}
\alias{k_to_f}
\title{Kelvin to fahrenheit conversion}
\usage{
k_to_f(x)
}
\arguments{
\item{x}{a vector containing a value in kelvin}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of temperature 
    from kelvin to fahrenheit
}
\examples{
k_to_f(75)
k_to_f(110)
k_to_f(0)
k_to_f(c(0, 75, 110))
}
