% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg_to_g.R
\name{kg_to_g}
\alias{kg_to_g}
\title{Kilogram to gram conversion}
\usage{
kg_to_g(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of weight
    from kilograms to grams
}
\examples{
kg_to_g(99)
kg_to_g(1033)
kg_to_g(687)
kg_to_g(c(350, 8332, 11033))
}
