% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg_to_t.R
\name{kg_to_t}
\alias{kg_to_t}
\title{Kilogram to ton conversion}
\usage{
kg_to_t(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of weight
    from kilograms to tons
}
\examples{
kg_to_t(99)
kg_to_t(1033)
kg_to_t(687)
kg_to_t(c(350, 8332, 11033))
}
