% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_to_mil.R
\name{km_to_mil}
\alias{km_to_mil}
\title{Kilometer to mile conversion}
\usage{
km_to_mil(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of length
    from kilometers to miles
}
\examples{
km_to_mil(75)
km_to_mil(110)
km_to_mil(0)
km_to_mil(c(0, 75, 110))
}
