% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_to_mm.R
\name{km_to_mm}
\alias{km_to_mm}
\title{Kilometer to milimeter conversion}
\usage{
km_to_mm(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of length
    from kilometers to milimeters
}
\examples{
km_to_mm(75)
km_to_mm(110)
km_to_mm(0)
km_to_mm(c(0, 75, 110))
}
