% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmh_to_ms.R
\name{kmh_to_ms}
\alias{kmh_to_ms}
\title{Kilometer per hour to meter per second conversion}
\usage{
kmh_to_ms(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of velocity
    from kilometers per hour to meters per second
}
\examples{
kmh_to_ms(75)
kmh_to_ms(110)
kmh_to_ms(0)
kmh_to_ms(c(0, 75, 110))
}
