% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_to_kb.R
\name{mb_to_kb}
\alias{mb_to_kb}
\title{Megabyte to kilobyte conversion}
\usage{
mb_to_kb(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of data size
    from megabytes to kilobytes
}
\examples{
mb_to_kb(99)
mb_to_kb(1033)
mb_to_kb(687)
mb_to_kb(c(350, 8332, 11033))
}
