% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_to_oz.R
\name{ml_to_oz}
\alias{ml_to_oz}
\title{Mililiter to ounce conversion}
\usage{
ml_to_oz(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of capacity
    from mililiters to ounces
}
\examples{
ml_to_oz(10)
ml_to_oz(325)
ml_to_oz(6876)
ml_to_oz(c(350, 8332, 11033))
}
