% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_to_kmh.R
\name{ms_to_kmh}
\alias{ms_to_kmh}
\title{Meter per second to kilometer per hour conversion}
\usage{
ms_to_kmh(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of velocity
    from meters per second to kilometers per hour
}
\examples{
ms_to_kmh(75)
ms_to_kmh(110)
ms_to_kmh(0)
ms_to_kmh(c(0, 75, 110))
}
