% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_to_bar.R
\name{pa_to_bar}
\alias{pa_to_bar}
\title{Pascal to bar conversion}
\usage{
pa_to_bar(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of pressure
    from pascals to bars
}
\examples{
pa_to_bar(99)
pa_to_bar(103)
pa_to_bar(67)
pa_to_bar(c(50, 80, 110))
}
