% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_to_kg.R
\name{t_to_kg}
\alias{t_to_kg}
\title{Ton to kilogram conversion}
\usage{
t_to_kg(x)
}
\arguments{
\item{x}{a vector containing a value greater than zero}
}
\value{
numeric vector
}
\description{
The function is used to convert a unit of weight
    from tons to kilograms
}
\examples{
t_to_kg(99)
t_to_kg(1033)
t_to_kg(687)
t_to_kg(c(350, 8332, 11033))
}
